+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Enhance Color Picker (react-color)
+
 ## [1.2.0] - 2016-03-02
 ### Added
 - Zoom